refactor(Data/Fin/Basic, Data/UInt)
Switch over to using the `[Inhabited (Fin n)]` bound, inilne lemmas where possible, and add USize. The Inhabited bound is needed to deal with the combination of `USize.size` not being defined in terms of `Nat.succ`, and `Zero` not giving enough information (by itself) to prove the lemmas needed for classes like AddMonoid. We need to know that `0` from OfNat.ofNat is definitionally equal to something that actually has the requisite properties.