feat: add withWeakNamespace (#151)
This is slightly different and potentially more useful than the mathport version, because it lets you escape the current namespace.  Which is what we want for the localized notation.  I've put in a version which appends the namespace for use in mathport (although I can't really tell what it's used for there