feat: LRAT proof replay
This is based on a suggestion by @avigad.
There is still some design work to be done on how exactly to make this
usable in practice; in particular the generated theorem is always
written in DNF, while for a tactic we would generally want to infer the
structure from the goal statement instead of the CNF file.
The main aim of this implementation is to be as high performance as
can be reasonably achieved while still being proof-object-based
(there is some definitional unfolding but nothing too heavy).
I haven't tried to give it any truly massive proofs yet
(I will have to figure out how to run it compiled first), but it seems
to hold up reasonably well even on proofs in the 10000 line range.