feat: Add support for building documentation via facets (#365)
This adds supports for building the documentation via a Lake facet as described in the README. The `meta if` is in place so that mathlib can still update a nightly version and have doc-gen breaking without its own build breaking in order for doc-gen not to becoome the critical path of a nightly update. Furthermore it also helps compile times as barely any users wants to actually use it.

The facet itself also has a huge performance advantage, on my local machine the old doc-gen4 from master with the unified (non parallel) build approach took 8-9 min to run. The new one that uses the full power of being parallelized by Lake on a per module basis only takes 2 minutes.

Furthermore once this PR is merged the CI of doc-gen4 should run through and deploy the first version that has (at least partial) support for linking back from the LeanInk rendered code into doc-gen pages.

Co-authored-by: Gabriel Ebner <gebner@gebner.org>