feat(Mathlib/Logic/Equiv/LocalEquiv): add the `mfld_set_tac` tactic (#331)
This PR adds the `mfld_set_tac` tactic. This tactic relies on a custom simp set which mostly contain lemmas about advanced math that is not yet in mathlib4. In the test file, we have created stubs of a random subset of the various definitions and lemmas it needs; this will need to be revisited during the port.

Co-authored-by: @hrmacbeth 

Co-authored-by: Heather Macbeth <25316162+hrmacbeth@users.noreply.github.com>
Co-authored-by: Scott Morrison <scott.morrison@gmail.com>