feat: log warning for unused names in `split_ifs with` (#877)
When `split_ifs with` splits a single if-then-else, it only consumes a single name (this matches the mathlib3 behavior). I noticed some places where we seem to be assuming that it consumes a separate name for each branch. To avoid that kind of confusion in the future, this PR adds a warning for any unused names.