feat: remove some @s and add some #aligns  (#1123)
This was supposed to be a review of #1001 but I missed the boat. Antoine added in a bunch of @s which were not present in the autoported file but were at that time necessary because MulOpposite was a type synonym. Now that it's a structure they're no longer needed so I removed them. Also a bunch of #aligns were missing related to to_additive.

Co-authored-by: Antoine Labelle <antoinelab01@gmail.com>
Co-authored-by: Scott Morrison <scott.morrison@gmail.com>
Co-authored-by: Johan Commelin <johan@commelin.net>