chore: remove superfluous imports from `runLinter` (#1322)
`Mathlib.Data.Array.Defs` is currently an empty file, because its contents have been upstreamed to Std, but it had `import Std` in it which means that building the linter requires compiling all of `Std`, which is not necessary. This is especially obvious now that builds are being cached since the only part that is not cached is the compilation of the C files.