fix: make push_neg try to synthesize LinearOrder before using it (#2095)
This fixes a bug in `push_neg` that occurs on expressions such as `¬(x ≤ y)` or `¬(x < y)`. Currently, it tries to transform these into `y < x` and `y ≤ x` even when `LinearOrder` is not present (and crashes if e.g. we only have a `Preorder`). This is fixed by making sure that we try to synthesize a `LinearOrder` before attempting the transformation.