fix: make convert tactic use `convert` suffix for goals from anonymous placeholders (#3314)
The `convert` tactic mistakenly calls `elabTermWithHoles` using the main goal's tag for the suffix. For example, if the main goal tag was `inl` this would produce tags like `inl.inl_1` for goals associated to anonymous placeholders. We change convert to use the suffix `convert` instead, yielding subgoals like `inl.convert_1` for these.

The previous behavior caused a panic in #3312 due to macro scopes in goal tags, which this fix happily sidesteps.