chore: copy over some nth_rewrite tests from mathlib3 (#3462)
There's a somewhat unsatisfactory situation here: when `nth_rewrite` was ported from mathlib3 to mathlib4, its behaviour changed significantly. Now it is "just" a syntax level tactic for accessing the "Occurrences" argument of the underlying rewrite machinery. However in mathlib3 is actually had different behaviour, allowing rewriting in individual locations even when the lemma matched in different ways. (Both Lean 3 and Lean 4 have a stuck metavariables problem in this situation.)

This PR restores the tests from mathlib3, with the ones broken because of this regression just commented out, and an explanation added. Hopefully one day we'll fix this up.



Co-authored-by: Scott Morrison <scott.morrison@gmail.com>