feat: pretty-print `Ordinal.{u}` (#5564)
Adds a `pp_with_univ` command that pretty-prints the universe parameters for the specified constant by default.  Just like `Type u`, we should include the universe parameter in `Ordinal.{u}` since you would otherwise have to guess it (and several lemmas require level inequalities, so it's important to know whether you have an `Ordinal.{u}` or merely an `Ordinal.{max u v}`).

It might be cute to pretty-print `Category.{v}` as well, but that is much more involved since the pretty-printer for constants is hardcoded.