feat(topology/order): Add closure.mono (#5631)
Adds `closure.mono` which asserts that if `t₁ ≤ t₂` for topologies `t₁`, `t₂`, then the closure of a set in `t₁` will be a subset of the closure of the set in `t₂`. Analogous to `IsOpen.mono` and `IsClosed.mono`. Also adds `closure[t₁]` notation to specify the topology to use.



Co-authored-by: Christopher Hoskin <mans0954@users.noreply.github.com>