feat: basic big operator plugin for `norm_num` (#4350)
I wrote a basic implementation for a `norm_num` plugin evaluating `Finset.sum` and `Finset.prod`. It does not yet support as many `Finset` expressions as the Mathlib 3 implementation, and it cannot deal with variables appearing in the sum. I believe it's valuable to have this implementation since we do use this plugin in various parts of mathlib.



Co-authored-by: Anne Baanen <Vierkantor@users.noreply.github.com>
Co-authored-by: Anne Baanen <t.baanen@vu.nl>
Co-authored-by: Jeremy Tan Jie Rui <reddeloostw@gmail.com>